Restructured how-to landing page
Previously, this was a bare list of sub-pages, not in any
discernible order, and hard to parse.

Now the sub-pages are grouped in sections by topic. It's
unlikely to be the final word on how this material is
arranged, but it's a clear improvement on the existing
arrangement and provides a good basis for next steps.